small fixes

+++ Engine/ac/string.h
@@ -19,10 +19,11 @@
 #define __AGS_EE_AC__STRING_H
 
 #include <stdarg.h>
+#include <stdint.h>
 #include "ac/dynobj/cc_dynamicobject.h"
 
 // Check that a supplied buffer from a text script function was not null
-#define VALIDATE_STRING(strin) if ((unsigned long)strin <= 4096) quit("!String argument was null: make sure you pass a string, not an int, as a buffer")
+#define VALIDATE_STRING(strin) if ((intptr_t)strin <= 4096) quit("!String argument was null: make sure you pass a string, not an int, as a buffer")
 
 int String_IsNullOrEmpty(const char *thisString);
 const char* String_Copy(const char *srcString);
